Cooperative ByteToMessageDecoder (WIP)
Motivation

This PR is very much a WIP but wanted to share more details of the idea
for feedback and discussion.

It grew out of the proposal from @belliotsmith in #8890 but is aimed at
more general applicability and ease of exploitation.

The core observation is that ByteToMessageDecoder is currently
unnecessarily inefficient w.r.t. how much data is moved around -
primarily the fact that the entire input buffer is copied (along with
any associated buffer reallocation costs) every time there is any
pre-existing data from a prior incomplete decode.

If the input buffer is very large it can be beneficial to first attempt
decoding with smaller copied amounts since this may allow the decoder
to advance beyond the pre-existing data, at which point we can switch to
use the remainder of the input buffer directly (rewinding first if
necessary).

Such an optimization can be implemented with no change to existing
decoders, but much greater efficiency is possible if we have basic
knowledge of how much input the particular decoder subclass requires at
a given point in time (instead of the current implicit signalling where
if the decoder does not increase the provided buffer's readerIndex then
it means more data is needed).

Here, a new final protected method setRequiredBytes(int min, int max) is
introduced which can optionally be called by subclasses prior to
returning from each decode() invocation, to provide information about
known data requirements for the _next_ invocation. The min and max args
specify necessary and sufficient byte counts for the decoder to make
progress (and will very often be equal). "Progress" here means either
some increase in the input buffer's reader index and/or acquisition of
enough new information to determine higher values for a subsequent call
to setRequiredBytes.

As described in #8890 there's also a benefit to cases where many inputs
are required to decode each message, since the cumulation buffer is
sized appropriately up-front meaning that it won't require
re-allocation/copying the same existing data each time an new buffer is
received.

Many existing decoders can exploit this and it's quick/easy to adapt
them (typically just adding a call to this method prior to each return,
with the args to pass in each case being straightforward to deduce). It
can also be called from the constructor which will correspond to the
_first_ call to decode(). A nice side effect is the potential for
simplification whereby decode() impls require fewer size checks since
it's guaranteed they will receive at minimum what was requested. It's
not immediately applicable to delimiter-based decoders but I have an
idea for an extension that could be made for those (not included yet).

IMHO for various reasons the current Cumulator interface boundary is not
the best mechanism for configuration/extensibility, given the purpose of
the handler. I think it's better to look at the current merge/composite
cumulator options as more of an optimization flag that decoders can set
if they don't require to be fed contiguous buffers, in turn based on the
kind of access that will be subsequently performed (random/granular
versus bulk). So I would advocate deprecating it in favour of something
like setRequiresContiguous(false). This could then even be signalled per
call to decode() similar to the size requirements in case only some
parts of the message require granular decoding.

Modifications

- Add min/maxRequired fields, protected setRequiredBytes(min,max) method
and single-arg convenience variant; incorporate reading/resetting of
these into callDecode(...) method logic
- Implement special-case logic for MERGE_CUMULATOR which takes the
min/max required values into account when aggregating data and
allocating new buffers
- Include special handling for case where the input buffer is composite,
treated as if the individual components were received sequentially. A
simple new componentSliceFromOffset(int) method was added to
CompositeByteBuf to facilitate this. (TODO: detect/handle _slices_ of
composite buffers similarly)
- Update ReplayingDecoder to automatically call setRequiredBytes() after
a failed read, but disable other direct use by subclasses for now
- Modify a few existing decoders to exploit the new capability
(currently
in separate commit): Http2ConnectionHandler, FixedLengthFrameDecoder,
LengthFieldBasedFrameDecoder, Lz4FrameDecoder,
ProtobufVarint32FrameDecoder, SslHandler

As part of the changes, already-read bytes are now never discarded from
the beginning of the cumulation buffer when MERGE_CUMULATOR is in use.

Still to be done (non-exhaustive):
- Benchmarks, more unit tests to exercise some relevant code paths not
currently covered
- More structural simplification and optimizations where possible
- Decide how best to expose the "never retain input buffer" flag (aims
to support this other specific requirement discussed in #8890)
- Consider adding an option to hold and reuse the cumulation buffer
instead of continual releasing/reallocating, maybe with a size cap. This
could greatly reduce pooled allocator contention under high concurrency
- Possible enhancement to the automatic ReplayableDecoder integration:
transitioning from the current cumulation to pending input could be done
if possible within checkpoint calls (reducing overall number of replays
required)
- Add support for delimiter-based codecs (such as LineBasedFrameDecoder,
HttpObjectDecoder)
- Enable more built-in decoders

Result

Hopefully reduced byte copying overhead on input path, especially when
data is large and/or network is slow. Benchmark results to follow.